guidance value number Defines the level of fidelity to the prompt when generating the image. A lower value allows for more creativity from the AI, while a higher value ensures closer adherence to the prompt. Valid values range [0.0, 2.0], default 1.0.
image.size
enum<string>
required
Image size with the aspect ratio. The aspect ratio is the proportional relationship between an image's width and height, expressed as *_width_height (e.g., square_1_1, widescreen_16_9). It is calculated by dividing the width by the height.

If not present, the default is square_1_1.
Note: For the fluid model, only this values are valid:

square_1_1
social_story_9_16
widescreen_16_9
traditional_3_4
classic_4_3